Cleaned house a little.  Changed wording on some things, and what I hope is better grammar
