+++ src/views/ApplicationViews.js
@@ -4,7 +4,7 @@ import { Register } from "../components/auth/Register"
 import { ViewAllCareTips } from "../components/careTips/allCareTips"
 import { CreateCareTip } from "../components/careTips/createCareTip"
 import { AllPlants } from "../components/plant/allPlants"
-import { AvailablePlants } from "../components/plant/availablePlants"
+import { AvailablePlants } from "../components/plant/viewAvailablePlants"
 import { CreateNewPlant } from "../components/plant/createNewPlant"
 import { EditCurrentPlant } from "../components/plant/editMyPlant"
 import { ViewMyPlants } from "../components/plant/myPlants"
